change consts to u64

+++ programs/spok/src/constants.rs
@@ -1,11 +1,11 @@
 pub const DECIMALS: u8 = 0;
 
 pub const MAX_TARGET: [u8; 32] = [255; 32];
-pub const MINTS_PER_TARGET_PERIOD: u8 = 1;
+pub const MINTS_PER_TARGET_PERIOD: u64 = 1;
 pub const SLOTS_PER_TARGET_PERIOD: u64 = 2;
 
 pub const INITIAL_SUBSIDY: u64 = 50;
-pub const MINTS_PER_SUBSIDY_PERIOD: u8 = 4;
+pub const MINTS_PER_SUBSIDY_PERIOD: u64 = 4;
 
 pub const MAX_TARGET_ADJ: f64 = 4.0;
 pub const MIN_TARGET_ADJ: f64 = 0.25;
